Actualización a registro de direcciones
Corregido id_ciudad a long en todas las clases usadas, cambios leves a la interfaz de registro (COLOR NARANJA E ICONOS QUE CAMBIAN DE COLOR!!!), cambios a checkbox y radio por los de mdb, boton de buscar actualizado (se retiro el marco negro que era muy molesto)

+++ src/java/control/CiudadDAO.java
@@ -32,7 +32,7 @@ public class CiudadDAO {
             try (PreparedStatement prepStmt = conexion.prepareStatement(strSQL)) {
                 ResultSet rs = prepStmt.executeQuery();
                 while (rs.next()) {
-                    ciudad.getId_ciudad_array().add(rs.getDouble(1));
+                    ciudad.getId_ciudad_array().add(rs.getLong(1));
                     ciudad.getNombre_array().add(rs.getString(2));
                     ciudad.getId_tienda_array().add(rs.getDouble(3));
                 }
